# DF/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/tob_lfi_consent/schema": {
      "get": {
        "description": "Retrive the datatype of the tob_lfi_consent table",
        "operationId": "tob_lfi_consentController_findSchema",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/tob_lfi_consentEntity"
                }
              }
            }
          }
        },
        "summary": "schema validation",
        "tags": [
          "ERD API"
        ]
      }
    },
    "/tob_lfi_consent/get": {
      "get": {
        "description": "Filter all the records from the tob_lfi_consent table",
        "operationId": "tob_lfi_consentController_findAllmethod",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of records to fetch",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/tob_lfi_consentEntity"
                  }
                }
              }
            }
          }
        },
        "summary": "Filter the records",
        "tags": [
          "ERD API"
        ]
      }
    },
    "/tob_lfi_consent/{id}": {
      "get": {
        "description": "Read all the records from the tob_lfi_consent table",
        "operationId": "tob_lfi_consentController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/tob_lfi_consentEntity"
                }
              }
            }
          }
        },
        "summary": "Read all the records",
        "tags": [
          "ERD API"
        ]
      },
      "patch": {
        "description": "Update the record for the tob_lfi_consent table",
        "operationId": "tob_lfi_consentController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Updatetob_lfi_consentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/tob_lfi_consentEntity"
                }
              }
            }
          }
        },
        "summary": "Update the record",
        "tags": [
          "ERD API"
        ]
      },
      "delete": {
        "description": "Delete the record for the tob_lfi_consent table",
        "operationId": "tob_lfi_consentController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/tob_lfi_consentEntity"
                }
              }
            }
          }
        },
        "summary": "Delete the record",
        "tags": [
          "ERD API"
        ]
      }
    },
    "/tob_lfi_consent/query": {
      "post": {
        "description": "Read all the records from the tob_lfi_consent table",
        "operationId": "tob_lfi_consentController_findAllwithquery",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Querytob_lfi_consentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/tob_lfi_consentEntity"
                  }
                }
              }
            }
          }
        },
        "summary": "Read all the records",
        "tags": [
          "ERD API"
        ]
      }
    },
    "/tob_lfi_consent": {
      "get": {
        "description": "Read all the records from the tob_lfi_consent table",
        "operationId": "tob_lfi_consentController_findAll",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/tob_lfi_consentEntity"
                  }
                }
              }
            }
          }
        },
        "summary": "Read all the records",
        "tags": [
          "ERD API"
        ]
      },
      "post": {
        "description": "Create the record for the tob_lfi_consent table",
        "operationId": "tob_lfi_consentController_create",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Createtob_lfi_consentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/tob_lfi_consentEntity"
                }
              }
            }
          }
        },
        "summary": "Create the record",
        "tags": [
          "ERD API"
        ]
      }
    },
...
  },
  "info": {
    "title": "Nest API",
    "description": "The Nestjs API documentation",
    "version": "0.1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://tgadev.gsstvl.com/ct242/tob001/tob002/v1/api-int",
      "description": "Production Server"
    }
  ],
  "components": {
    "schemas": {
...
      "tob_lfi_dataEntity": {
        "type": "object",
        "properties": {
          "ids": {
            "type": "string"
          },
          "parid": {
            "type": "string"
          },
          "rartype": {
            "type": "string"
          },
          "standardversion": {
            "type": "string"
          },
          "consentgroupid": {
            "type": "string"
          },
          "requesturl": {
            "type": "string"
          },
          "consenttype": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "consentid": {
            "type": "string"
          }
        },
        "required": [
          "ids",
          "parid",
          "rartype",
          "standardversion",
          "consentgroupid",
          "requesturl",
          "consenttype",
          "status",
          "consentid"
        ]
      },
      "tob_lfi_consentEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/tob_lfi_dataEntity"
          },
          "trs_created_date": {
            "format": "date-time",
            "type": "string"
          },
          "trs_created_by": {
            "type": "string"
          },
          "trs_modified_date": {
            "format": "date-time",
            "type": "string"
          },
          "trs_modified_by": {
            "type": "string"
          },
          "trs_status": {
            "type": "string"
          },
          "trs_next_status": {
            "type": "string"
          },
          "trs_process_id": {
            "type": "string"
          },
          "trs_access_profile": {
            "type": "string"
          },
          "trs_org_grp_code": {
            "type": "string"
          },
          "trs_org_code": {
            "type": "string"
          },
          "trs_role_grp_code": {
            "type": "string"
          },
          "trs_role_code": {
            "type": "string"
          },
          "trs_ps_grp_code": {
            "type": "string"
          },
          "trs_ps_code": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "data",
          "trs_created_date",
          "trs_created_by",
          "trs_modified_date",
          "trs_modified_by",
          "trs_status",
          "trs_next_status",
          "trs_process_id",
          "trs_access_profile",
          "trs_org_grp_code",
          "trs_org_code",
          "trs_role_grp_code",
          "trs_role_code",
          "trs_ps_grp_code",
          "trs_ps_code"
        ]
      },
      "Querytob_lfi_dataDto": {
        "type": "object",
        "properties": {
          "ids": {
            "type": "string"
          },
          "parid": {
            "type": "string"
          },
          "rartype": {
            "type": "string"
          },
          "standardversion": {
            "type": "string"
          },
          "consentgroupid": {
            "type": "string"
          },
          "requesturl": {
            "type": "string"
          },
          "consenttype": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "consentid": {
            "type": "string"
          }
        },
        "required": [
          "ids",
          "parid",
          "rartype",
          "standardversion",
          "consentgroupid",
          "requesturl",
          "consenttype",
          "status",
          "consentid"
        ]
      },
      "Querytob_lfi_dataWapperDto": {
        "type": "object",
        "properties": {
          "is": {
            "$ref": "#/components/schemas/Querytob_lfi_dataDto"
          }
        },
        "required": [
          "is"
        ]
      },
      "Querytob_lfi_consentDto": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Querytob_lfi_dataWapperDto"
          },
          "trs_created_date": {
            "type": "string",
            "format": "date-time"
          },
          "trs_created_by": {
            "type": "string"
          },
          "trs_modified_date": {
            "type": "string",
            "format": "date-time"
          },
          "trs_modified_by": {
            "type": "string"
          },
          "trs_status": {
            "type": "string"
          },
          "trs_next_status": {
            "type": "string"
          },
          "trs_process_id": {
            "type": "string"
          },
          "trs_access_profile": {
            "type": "string"
          },
          "trs_org_grp_code": {
            "type": "string"
          },
          "trs_org_code": {
            "type": "string"
          },
          "trs_role_grp_code": {
            "type": "string"
          },
          "trs_role_code": {
            "type": "string"
          },
          "trs_ps_grp_code": {
            "type": "string"
          },
          "trs_ps_code": {
            "type": "string"
          }
        },
        "required": [
          "data",
          "trs_created_date",
          "trs_created_by",
          "trs_modified_date",
          "trs_modified_by",
          "trs_status",
          "trs_next_status",
          "trs_process_id",
          "trs_access_profile",
          "trs_org_grp_code",
          "trs_org_code",
          "trs_role_grp_code",
          "trs_role_code",
          "trs_ps_grp_code",
          "trs_ps_code"
        ]
      },
      "Createtob_lfi_dataDto": {
        "type": "object",
        "properties": {
          "ids": {
            "type": "string"
          },
          "parid": {
            "type": "string"
          },
          "rartype": {
            "type": "string"
          },
          "standardversion": {
            "type": "string"
          },
          "consentgroupid": {
            "type": "string"
          },
          "requesturl": {
            "type": "string"
          },
          "consenttype": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "consentid": {
            "type": "string"
          }
        },
        "required": [
          "ids",
          "parid",
          "rartype",
          "standardversion",
          "consentgroupid",
          "requesturl",
          "consenttype",
          "status",
          "consentid"
        ]
      },
      "Createtob_lfi_consentDto": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Createtob_lfi_dataDto"
          },
          "trs_created_date": {
            "type": "string",
            "format": "date-time"
          },
          "trs_created_by": {
            "type": "string"
          },
          "trs_modified_date": {
            "type": "string",
            "format": "date-time"
          },
          "trs_modified_by": {
            "type": "string"
          },
          "trs_status": {
            "type": "string"
          },
          "trs_next_status": {
            "type": "string"
          },
          "trs_process_id": {
            "type": "string"
          },
          "trs_access_profile": {
            "type": "string"
          },
          "trs_org_grp_code": {
            "type": "string"
          },
          "trs_org_code": {
            "type": "string"
          },
          "trs_role_grp_code": {
            "type": "string"
          },
          "trs_role_code": {
            "type": "string"
          },
          "trs_ps_grp_code": {
            "type": "string"
          },
          "trs_ps_code": {
            "type": "string"
          }
        },
        "required": [
          "data",
          "trs_created_date",
          "trs_created_by",
          "trs_modified_date",
          "trs_modified_by",
          "trs_status",
          "trs_next_status",
          "trs_process_id",
          "trs_access_profile",
          "trs_org_grp_code",
          "trs_org_code",
          "trs_role_grp_code",
          "trs_role_code",
          "trs_ps_grp_code",
          "trs_ps_code"
        ]
      },
      "Updatetob_lfi_dataDto": {
        "type": "object",
        "properties": {
          "ids": {
            "type": "string"
          },
          "parid": {
            "type": "string"
          },
          "rartype": {
            "type": "string"
          },
          "standardversion": {
            "type": "string"
          },
          "consentgroupid": {
            "type": "string"
          },
          "requesturl": {
            "type": "string"
          },
          "consenttype": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "consentid": {
            "type": "string"
          }
        },
        "required": [
          "ids",
          "parid",
          "rartype",
          "standardversion",
          "consentgroupid",
          "requesturl",
          "consenttype",
          "status",
          "consentid"
        ]
      },
      "Updatetob_lfi_consentDto": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Updatetob_lfi_dataDto"
          },
          "trs_created_date": {
            "type": "string",
            "format": "date-time"
          },
          "trs_created_by": {
            "type": "string"
          },
          "trs_modified_date": {
            "type": "string",
            "format": "date-time"
          },
          "trs_modified_by": {
            "type": "string"
          },
          "trs_status": {
            "type": "string"
          },
          "trs_next_status": {
            "type": "string"
          },
          "trs_process_id": {
            "type": "string"
          },
          "trs_access_profile": {
            "type": "string"
          },
          "trs_org_grp_code": {
            "type": "string"
          },
          "trs_org_code": {
            "type": "string"
          },
          "trs_role_grp_code": {
            "type": "string"
          },
          "trs_role_code": {
            "type": "string"
          },
          "trs_ps_grp_code": {
            "type": "string"
          },
          "trs_ps_code": {
            "type": "string"
          }
        },
        "required": [
          "data",
          "trs_created_date",
          "trs_created_by",
          "trs_modified_date",
          "trs_modified_by",
          "trs_status",
          "trs_next_status",
          "trs_process_id",
          "trs_access_profile",
          "trs_org_grp_code",
          "trs_org_code",
          "trs_role_grp_code",
          "trs_role_code",
          "trs_ps_grp_code",
          "trs_ps_code"
        ]
      },
...
    }
  }
}